feat: UI and delete
